스트릭 api 기능 운영 환경에 배포

+++ src/main/java/com/yeo_li/yeol_post/post/controller/PostController.java
@@ -105,7 +105,7 @@ public class PostController {
     }
 
     @DeleteMapping("/{postId}")
-    public ResponseEntity<?> deletePost(@PathVariable("postId") int postId) {
+    public ResponseEntity<?> deletePost(@PathVariable("postId") Long postId) {
 
         postService.deletePostByPostId(postId);
